+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/storage/annotation/Column.java
@@ -49,20 +49,13 @@ public @interface Column {
     int defaultValue() default 0;
 
     /**
-     * Match query means using analyzer(if storage have) to do key word match query.
-     */
-    boolean matchQuery() default false;
-
-    /**
-     * The column is just saved, never used in query.
+     * The column is just saved, never used as a query condition.
      */
     boolean storageOnly() default false;
 
     /**
-     * The column(field) is just indexed, never stored. Note: this feature only supported by elasticsearch and don't
-     * support mappings update due to ElasticSearch server's limitation.
-     *
-     * NOTICE, metrics should not use this, as the OAP core merges indices of metrics automatically.
+     * The column(field) is just indexed, never stored(not available in query projection).
+     * Note: this feature only supported by elasticsearch.
      */
     boolean indexOnly() default false;
 
@@ -71,7 +64,7 @@ public @interface Column {
      * storage implementation.
      *
      * Notice, different lengths may cause different types. Such as, over 16383 would make the type in MySQL to be
-     * MEDIUMTEXT, due to database varchar max=16383
+     * MEDIUMTEXT, due to database varchar max=16383.
      * @since 7.1.0
      */
     int length() default 200;
@@ -93,57 +86,6 @@ public @interface Column {
      * @since 8.0.0
      */
     ValueDataType dataType() default ValueDataType.NOT_VALUE;
-
-    /**
-     * The storage analyzer mode.
-     *
-     * @since 8.4.0
-     */
-    AnalyzerType analyzer() default AnalyzerType.OAP_ANALYZER;
-
-    /**
-     * Sharding key is used to group time series data per metric of one entity in one place (same sharding and/or same
-     * row for column-oriented database).
-     * For example,
-     * ServiceA's traffic gauge, service call per minute, includes following timestamp values, then it should be sharded
-     * by service ID
-     * [ServiceA(encoded ID): 01-28 18:30 values-1, 01-28 18:31 values-2, 01-28 18:32 values-3, 01-28 18:32 values-4]
-     *
-     * BanyanDB is the 1st storage implementation supporting this. It would make continuous time series metrics stored
-     * closely and compressed better.
-     *
-     * 1. One entity could have multiple sharding keys
-     * 2. If no column is appointed for this, {@link org.apache.skywalking.oap.server.core.storage.StorageData#id}
-     * would be used by the storage implementation accordingly.
-     *
-     * NOTICE, this sharding concept is NOT just for splitting data into different database instances or physical
-     * files.
-     *
-     * @return non-negative if this column be used for sharding. -1 means not as a sharding key
-     * @since 9.0.0
-     */
-    int shardingKeyIdx() default -1;
-
-    /**
-     * The analyzer declares the text analysis mode.
-     */
-    enum AnalyzerType {
-        /**
-         * The default analyzer.
-         */
-        OAP_ANALYZER("oap_analyzer"),
-        /**
-         * The log analyzer.
-         */
-        OAP_LOG_ANALYZER("oap_log_analyzer");
-
-        @Getter
-        private final String name;
-
-        AnalyzerType(final String name) {
-            this.name = name;
-        }
-    }
 
     /**
      * ValueDataType represents the data structure of value column. The persistent way of the value column determine the